added code for checking due date and send email
