+++ GuiltMotivator/app/src/main/java/nielsen/guiltmotivator/NotificationIntentService.java
@@ -25,15 +25,13 @@ public class NotificationIntentService extends IntentService {
     private static final String ACTION_START = "ACTION_START";
     private static final String ACTION_DELETE = "ACTION_DELETE";
     private boolean isEmailSent = false;
-    private static WeakReference<Activity> mActivityRef;
+
 
     public NotificationIntentService() {
         super(NotificationIntentService.class.getSimpleName());
     }
 
-    public static void updateActivity(Activity activity) {
-        mActivityRef = new WeakReference<>(activity);
-    }
+
 
     public static Intent createIntentStartNotificationService(Context context) {
         
@@ -95,37 +93,6 @@ public class NotificationIntentService extends IntentService {
 
         final NotificationManager manager = (NotificationManager) this.getSystemService(Context.NOTIFICATION_SERVICE);
         manager.notify(m, builder.build());
-
-        //sendEmail();
-    }
-
-
-    private void sendEmail(){
-//        Activity activity = (Activity) context;
-        //get helper and get db in write mode
-        DatabaseHelper mDbHelper = new DatabaseHelper(getApplicationContext());
-        final SQLiteDatabase db = mDbHelper.getWritableDatabase();
-        ArrayList<Task> list = mDbHelper.getAllTasks();
-        Task task = list.get(0);
-        ArrayList<Contact> contacts = mDbHelper.getContacts(task);
-
-        Activity activity = mActivityRef.get();
-
-        final SharedPreferences sharedPref = activity.getPreferences(Context.MODE_PRIVATE);
-        final String tone = sharedPref.getString(MainActivity.SAVED_TONE, "polite");
-        final String name = sharedPref.getString(MainActivity.SAVED_NAME, "none");
-        final String politeMsg = "polite";
-        final String profaneMsg = "profane";
-//        String msg = tone == "polite"? politeMsg : profaneMsg;
-        String msg = name + tone;
-        if (!task.isChecked()){
-            for (int i = 0; i < contacts.size();i++){
-                //Creating SendMail object
-                SendMail sm = new SendMail(this, contacts.get(i).getAddress(), "From Guilt Motivator", msg);
-                //Executing sendmail to send email
-                sm.execute();
-            }
-        }
     }
 
 }
